livekit_agent/requirements-livekit.txt
livekit-agents[openai]~=1.4
livekit-agents[deepgram]~=1.4
livekit-agents[silero]~=1.4
livekit-api
livekit-plugins-simli
python-dotenv
